Combine typings
